+++ templates/index.html
@@ -194,7 +194,7 @@
   </div>
 
   <footer>
-    制作：<a href="https://space.bilibili.com/1969358629?spm_id_from=333.1007.0.0" target="_blank">哎哎哎哎哎哎哎哎哎啦</a>、<a href="https://github.com/neyuki778" target="_blank">Yuki</a>、<a href="https://tieba.baidu.com/home/main?id=tb.1.69a710ab.Wys0SW7-Na-4ZOdVt0XGgg" target="_blank">平者深黑</a>、<a href="https://github.com/jiaweiz9" target="_blank">Kyle</a>、<a href="https://github.com/Pseudocirrus" target="_blank">Pseudocirrus</a>、<a href="https://space.bilibili.com/443610657" target="_blank">之徙zx</a>、[虚位以待]、 [虚位以待] 等
+    制作：<a href="https://space.bilibili.com/1969358629?spm_id_from=333.1007.0.0" target="_blank">哎哎哎哎哎哎哎哎哎啦</a>、Yuki、<a href="https://tieba.baidu.com/home/main?id=tb.1.69a710ab.Wys0SW7-Na-4ZOdVt0XGgg" target="_blank">平者深黑</a>、Kyle Zhang、Pseudocirrus、<a href="https://space.bilibili.com/443610657" target="_blank">之徙zx</a>、[虚位以待]、 [虚位以待] 等
   </footer>
 
   <div class="tab-bar">
